separate ts config
