+++ packages/insomnia/src/main/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["./**/*"],
+  "exclude": ["../../types/renderer-global.d.ts"]
+}
