+++ packages/insomnia/tsconfig.json
@@ -33,6 +33,7 @@
   },
   "include": ["**/*", ".react-router/types/**/*"],
   "exclude": [
+    "./src/main",
     "./src/__mocks__/@getinsomnia/node-libcurl.ts",
     "**/__mocks__",
     "**/__snapshots__",
